+++ Code/IO/btkANxFileIOUtils_p.h
@@ -58,11 +58,11 @@ namespace btk
     virtual ~ANxFileIOException() throw() {};
   };
   
-  void ANxFileIOCheckHeader_p(double preciseRate, int channelNumber, 
+  void ANxFileIOCheckHeader_p(double preciseRate, size_t channelNumber, 
                             const std::vector<uint16_t>& channelRate, 
                             const std::vector<uint16_t>& channelRange);
   void ANxFileIOStoreHeader_p(Acquisition::Pointer output,
-                            double preciseRate, int frameNumber, int channelNumber,
+                            double preciseRate, size_t frameNumber, size_t channelNumber,
                             const std::vector<std::string>& channelLabel,
                             const std::vector<uint16_t>& channelRate,
                             const std::vector<uint16_t>& channelRange,
